backend/requirements.txt
python-dotenv>=1.0.1
pydantic>=2.8.2
pandas>=2.2.2
openpyxl>=3.1.5
faiss-cpu>=1.8.0.post1
sentence-transformers>=3.0.1
transformers>=4.44.2
selectolax>=0.3.21
playwright>=1.47.0
fastapi>=0.104.0
uvicorn>=0.24.0
groq>=0.9.0
httpx>=0.27.0
urllib3>=2.2.2
beautifulsoup4>=4.12.3
requests>=2.31.0
